requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
beautifulsoup4==4.8.1
cffi==1.13.2
chardet==2.3.0
constantly==15.1.0
cryptography==2.8
cssutils==1.0.2
daphne==2.3.0
Django==2.2.7
django-cors-headers==3.1.0
django-debug-toolbar==2.0
django-inlinecss==0.3.0
django-rest-passwordreset==1.0.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
future==0.18.2
hyperlink==19.0.0
idna==2.2
incremental==17.5.0
jeepney==0.4.1
keyring==10.1
keyrings.alt==1.3
olefile==0.46
pexpect==4.2.1
Pillow==4.0.0
psycopg2-binary==2.8.3
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.19
pycrypto==2.6.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pynliner==0.8.0
pytz==2019.3
pyxdg==0.25
reportlab==3.3.0
requests==2.12.4
SecretStorage==2.3.1
six==1.10.0
soupsieve==1.9.5
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.19.1
zope.interface==4.7.1
